Performance optimisation: use base*base*base instead of pow(base,3)
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@1150 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ include/WithinHost/Infection/MolineauxInfection.h
@@ -68,13 +68,12 @@ private:
 	static double mean_diff_pos_days, sd_diff_pos_days;
 	//@}
 	
-	//@brief q^(i+1) array
-	//@{
-	/// all the values of q^1... q^50 are stored in this array.
-	/// this prevent the recalculation of those values on every two timesteps.
+	/** @brief q^(i+1) array
+	 *
+	 * all the values of q^1... q^50 are stored in this array.
+	 * this prevent the recalculation of those values on every two timesteps. */
 	static double qPow[v];
-	//@}
-
+	
 	/** @brief The static variables (double)
 	 *
 	 * sProb: fraction of parasites switching among variants per two-day cycle
@@ -98,10 +97,11 @@ private:
 	static const double k_c=0.2;
 	static const double k_m=0.04;
 	static const double Pstar_v=30.0;
-	static const double kappa_c=3.0;
-	static const double kappa_m=1.0;
-	static const double kappa_v=3.0;
+	static const int kappa_c=3;
+	static const int kappa_m=1;
+	static const int kappa_v=3;
 	static const double C=1.0;
+	//@}
 
 	/*
 	 * The dynamic variables:
@@ -129,6 +129,7 @@ private:
 	    double updateDensity (double survivalFactor, int ageOfInfection);
 	    double getVariantSpecificSummation();
 	};
+	//TODO: resizes may be slow; should check.
 	vector<Variant> variants;
 };
